Adding groundschool logic

+++ src/main/java/com/starfireaviation/groundschool/users/UserCreatedEvent.java
@@ -3,5 +3,5 @@ package com.starfireaviation.groundschool.users;
 import org.springframework.modulith.events.Externalized;
 
 @Externalized(target = AmqpIntegrationConfiguration.USERS_Q)
-public record UserCreatedEvent(int id) {
+public record UserCreatedEvent(Long id) {
 }
